refactor: share public journey primitives
